requirements.txt
requests
python-telegram-bot
redis